+++ api/Dtos/Discount/DiscountDto.cs
@@ -7,7 +7,7 @@ namespace api.Dtos.Discount
         public int Id { get; set; }
         public int MerchantId { get; set; }
         public string Title { get; set; } = string.Empty;
-        public decimal Percentage { get; set; }
+        public int Percentage { get; set; }
         public DateTime ExpiresOn { get; set; }
         public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
         public DateTime UpdatedAt { get; set; } = DateTime.UtcNow;
